move program filtering logic into its own file
